+++ src/main/java/com/buuz135/replication/api/matter_fluid/client/MatterTankScreenAddon.java
@@ -46,7 +46,7 @@ public class MatterTankScreenAddon extends BasicScreenAddon {
         Rectangle area = asset.getArea();
         if (!tank.getMatter().isEmpty()) {
             MatterStack matterStack = tank.getMatter();
-            double stored = tank.getMatterAmount();
+            double stored = Math.min(tank.getMatterAmount(), tank.getCapacity());
             double capacity = tank.getCapacity();
             int topBottomPadding = asset.getFluidRenderPadding(Direction.UP) + asset.getFluidRenderPadding(Direction.DOWN);
             int offset = (int) ((stored / capacity) * (area.height - topBottomPadding));
@@ -84,7 +84,7 @@ public class MatterTankScreenAddon extends BasicScreenAddon {
     public java.util.List<net.minecraft.network.chat.Component> getTooltipLines() {
         List<Component> strings = new ArrayList<>();
         strings.add(Component.literal(ChatFormatting.GOLD + Component.translatable("tooltip.replication.tank.matter").getString()).append(tank.getMatter().isEmpty() ? Component.translatable("tooltip.titanium.tank.empty").withStyle(ChatFormatting.WHITE) : Component.translatable(tank.getMatter().getTranslationKey())).withStyle(ChatFormatting.WHITE));
-        strings.add(net.minecraft.network.chat.Component.translatable("tooltip.replication.terminal.amount").withStyle(ChatFormatting.GOLD).append(Component.literal(ChatFormatting.WHITE + new DecimalFormat().format(tank.getMatterAmount()) + ChatFormatting.GOLD + "/" + ChatFormatting.WHITE + new DecimalFormat().format(tank.getCapacity()))).append(Component.translatable("tooltip.replication.tank.unit").withStyle(ChatFormatting.DARK_AQUA)));
+        strings.add(net.minecraft.network.chat.Component.translatable("tooltip.replication.terminal.amount").withStyle(ChatFormatting.GOLD).append(Component.literal(ChatFormatting.WHITE + new DecimalFormat().format(tank.getMatter().getAmount()) + ChatFormatting.GOLD + "/" + ChatFormatting.WHITE + new DecimalFormat().format(tank.getCapacity()))).append(Component.translatable("tooltip.replication.tank.unit").withStyle(ChatFormatting.DARK_AQUA)));
         return strings;
     }