+++ src/main/java/com/buuz135/replication/api/network/IMatterTanksConsumer.java
@@ -8,4 +8,6 @@ public interface IMatterTanksConsumer {
 
     List<? extends IMatterTank> getTanks();
 
+    public int getPriority();
+
 }