+++ src/main/java/com/buuz135/replication/api/network/IMatterTanksSupplier.java
@@ -8,4 +8,6 @@ public interface IMatterTanksSupplier {
 
     List<? extends IMatterTank> getTanks();
 
+    public int getPriority();
+
 }