+++ src/main/java/com/buuz135/replication/compat/JadePlugin.java
@@ -1,8 +1,8 @@
 package com.buuz135.replication.compat;
 
-import com.buuz135.replication.api.matter_fluid.MatterStack;
+import com.buuz135.replication.block.CreativeMatterTankBlock;
 import com.buuz135.replication.block.MatterTankBlock;
-import com.buuz135.replication.block.tile.MatterTankBlockEntity;
+import com.buuz135.replication.block.tile.BaseMatterTankBlockEntity;
 import com.buuz135.replication.compat.jade.MatterTankComponentProvider;
 import snownee.jade.api.IWailaClientRegistration;
 import snownee.jade.api.IWailaCommonRegistration;
@@ -16,12 +16,12 @@ public class JadePlugin implements IWailaPlugin {
 
     @Override
     public void register(IWailaCommonRegistration registration) {
-        registration.registerBlockDataProvider(componentProvider, MatterTankBlockEntity.class);
+        registration.registerBlockDataProvider(componentProvider, BaseMatterTankBlockEntity.class);
     }
 
     @Override
     public void registerClient(IWailaClientRegistration registration) {
         registration.registerBlockComponent(componentProvider, MatterTankBlock.class);
-
+        registration.registerBlockComponent(componentProvider, CreativeMatterTankBlock.class);
     }
 }